pyproject.toml
[project]
name = "projeto-agno-seo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=2.5.5",
    "chainlit>=2.9.6",
    "ddgs>=9.10.0",
    "fastapi>=0.133.1",
    "google-genai>=1.65.0",
    "groq>=1.0.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
]
